+++ src/index.d.ts
@@ -8,7 +8,16 @@ import {
 import {
 	addItemToPlaylist,
 	createPlaylist,
-	fetchUserPlaylists
+	deleteItemsFromPlaylist,
+	duplicatePlaylist,
+	fetchPlaylist,
+	fetchPlaylistItems,
+	fetchUserPlaylists,
+	likePlaylist,
+	reportPlaylist,
+	togglePlaylistPrivate,
+	unlikePlaylist,
+	updatePlaylist
 } from './services/content-org/playlists.js';
 
 import {
@@ -22,7 +31,12 @@ import {
 } from './services/content.js';
 
 import {
+	addContextToContent
+} from './services/contentAggregator.js';
+
+import {
 	isContentLiked,
+	isContentLikedByIds,
 	likeContent,
 	unlikeContent
 } from './services/contentLikes.js';
@@ -40,6 +54,7 @@ import {
 	getProgressState,
 	getProgressStateByIds,
 	getResumeTimeSeconds,
+	getResumeTimeSecondsByIds,
 	recordWatchSession
 } from './services/contentProgress.js';
 
@@ -78,13 +93,8 @@ import {
 import {
 	assignModeratorToComment,
 	closeComment,
-	countAssignmentsAndLessons,
 	createComment,
 	deleteComment,
-	deletePlaylist,
-	deletePlaylistItem,
-	deletePlaylistLike,
-	duplicatePlaylist,
 	editComment,
 	fetchAllCompletedStates,
 	fetchCarouselCardData,
@@ -101,12 +111,10 @@ import {
 	fetchContentPageUserData,
 	fetchContentProgress,
 	fetchHandler,
+	fetchLastInteractedChild,
+	fetchLikeCount,
 	fetchNextContentDataForParent,
 	fetchOwnedChallenges,
-	fetchPinnedPlaylists,
-	fetchPlaylist,
-	fetchPlaylistItem,
-	fetchPlaylistItems,
 	fetchSongsInProgress,
 	fetchTopComment,
 	fetchUserAward,
@@ -118,11 +126,8 @@ import {
 	fetchUserPracticeNotes,
 	fetchUserPractices,
 	likeComment,
-	likePlaylist,
 	logUserPractice,
 	openComment,
-	pinPlaylist,
-	playback,
 	postChallengesCommunityNotification,
 	postChallengesEnroll,
 	postChallengesEnrollmentNotification,
@@ -138,13 +143,9 @@ import {
 	postRecordWatchSession,
 	replyToComment,
 	reportComment,
-	reportPlaylist,
 	setStudentViewForUser,
 	unassignModeratorToComment,
-	unlikeComment,
-	unpinPlaylist,
-	updatePlaylist,
-	updatePlaylistItem
+	unlikeComment
 } from './services/railcontent.js';
 
 import {
@@ -220,13 +221,13 @@ import {
 } from './services/user/sessions.js';
 
 import {
+	calculateLongestStreaks,
 	createPracticeNotes,
 	deletePracticeSession,
 	getPracticeNotes,
 	getPracticeSessions,
 	getRecentActivity,
 	getUserMonthlyStats,
-	getUserPractices,
 	getUserWeeklyStats,
 	recordUserPractice,
 	removeUserPractice,
@@ -235,8 +236,6 @@ import {
 	updatePracticeNotes,
 	updateUserPractice
 } from './services/userActivity.js';
-
-import { addContextToContent } from './contentAggregator.js';
 
 declare module 'musora-content-services' {
 	export {
@@ -249,18 +248,16 @@ declare module 'musora-content-services' {
 		assignmentStatusReset,
 		blockUser,
 		buildImageSRC,
+		calculateLongestStreaks,
 		closeComment,
 		contentStatusCompleted,
 		contentStatusReset,
 		convertToTimeZone,
-		countAssignmentsAndLessons,
 		createComment,
 		createPlaylist,
 		createPracticeNotes,
 		deleteComment,
-		deletePlaylist,
-		deletePlaylistItem,
-		deletePlaylistLike,
+		deleteItemsFromPlaylist,
 		deletePracticeSession,
 		duplicatePlaylist,
 		editComment,
@@ -296,9 +293,11 @@ declare module 'musora-content-services' {
 		fetchGenreLessons,
 		fetchHandler,
 		fetchHierarchy,
+		fetchLastInteractedChild,
 		fetchLeaving,
 		fetchLessonContent,
 		fetchLessonsFeaturingThisContent,
+		fetchLikeCount,
 		fetchLiveEvent,
 		fetchMetadata,
 		fetchMethod,
@@ -313,10 +312,8 @@ declare module 'musora-content-services' {
 		fetchPackAll,
 		fetchPackData,
 		fetchParentForDownload,
-		fetchPinnedPlaylists,
 		fetchPlayAlongsCount,
 		fetchPlaylist,
-		fetchPlaylistItem,
 		fetchPlaylistItems,
 		fetchRecent,
 		fetchRelatedLessons,
@@ -363,17 +360,18 @@ declare module 'musora-content-services' {
 		getRecentActivity,
 		getRecommendedForYou,
 		getResumeTimeSeconds,
+		getResumeTimeSecondsByIds,
 		getScheduleContentRows,
 		getSortOrder,
 		getTabResults,
 		getUserMonthlyStats,
-		getUserPractices,
 		getUserWeeklyStats,
 		getWeekNumber,
 		globalConfig,
 		initializeService,
 		isBucketUrl,
 		isContentLiked,
+		isContentLikedByIds,
 		isNextDay,
 		isSameDate,
 		jumpToContinueContent,
@@ -384,8 +382,6 @@ declare module 'musora-content-services' {
 		login,
 		logout,
 		openComment,
-		pinPlaylist,
-		playback,
 		postChallengesCommunityNotification,
 		postChallengesEnroll,
 		postChallengesEnrollmentNotification,
@@ -412,13 +408,13 @@ declare module 'musora-content-services' {
 		restorePracticeSession,
 		restoreUserPractice,
 		setStudentViewForUser,
+		togglePlaylistPrivate,
 		unassignModeratorToComment,
 		unblockUser,
 		unlikeComment,
 		unlikeContent,
-		unpinPlaylist,
+		unlikePlaylist,
 		updatePlaylist,
-		updatePlaylistItem,
 		updatePracticeNotes,
 		updateUserPractice,
 		verifyImageSRC,